requirements.txt
astropy
mypy
pyserial
PyOpenGL
PyOpenGL_accelerate
PyOpenGL-Demo
pyyaml
scipy
sgp4
